pyproject.toml
[tool.poetry]
name = "document-chat"
version = "0.1.0"
description = "A system for ingesting documents and providing interactive chat on their contents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.25"
openai = "^1.86.0"
faiss-cpu = "^1.11.0"
pypdf = "^5.6.0"
python-docx = "^1.1.2"
typer = "^0.16.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
